frontend/requirements.txt
streamlit==1.42.0
requests
PyPDF2
python-docx 